+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add Asterisk to Select step
 - Add selectAsteriskFrom function to Builder
 - Add Steps RailRoad in doc
+- Add OrderBy step
 
 ## [0.4.9]  2022-01-27
 ### Fixed